Issue #4587: Moved all resources for SuppressWarnings check
