+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/SuppressWarningsHolderTest.java
@@ -56,12 +56,18 @@ import com.puppycrawl.tools.checkstyle.utils.CommonUtils;
 public class SuppressWarningsHolderTest extends BaseCheckTestSupport {
     @Override
     protected String getPath(String filename) throws IOException {
-        return super.getPath("checks" + File.separator + filename);
+        return super.getPath("checks" + File.separator
+                + "misc" + File.separator
+                + "suppresswarnings" + File.separator
+                + filename);
     }
 
     @Override
     protected String getNonCompilablePath(String filename) throws IOException {
-        return super.getNonCompilablePath("checks" + File.separator + filename);
+        return super.getNonCompilablePath("checks" + File.separator
+                + "misc" + File.separator
+                + "suppresswarnings" + File.separator
+                + filename);
     }
 
     @Test
