--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/InputSuppressWarningsHolder4.java
@@ -1,10 +1,0 @@
-package com.puppycrawl.tools.checkstyle.checks;
-
-public class InputSuppressWarningsHolder4 {
-
-    @com.puppycrawl.tools.checkstyle.checks.CustomAnnotation()
-    int a;
-    
-}
-@interface CustomAnnotation {
-}
