+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/suppresswarnings/InputSuppressWarningsHolder2.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks;
+package com.puppycrawl.tools.checkstyle.checks.misc.suppresswarnings;
 
 public class InputSuppressWarningsHolder2
 {
