+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/suppresswarnings/InputSuppressWarningsHolder4.java
@@ -1,0 +1,10 @@
+package com.puppycrawl.tools.checkstyle.checks.misc.suppresswarnings;
+
+public class InputSuppressWarningsHolder4 {
+
+    @CustomAnnotation()
+    int a;
+    
+}
+@interface CustomAnnotation {
+}
